Update to Version 0.3.0
This update re-incorporates the GameAnalytics plugin and modifies the system to be able to send to all analytic endpoints at once.
To make this update work you will need to download and install the following Unity Package file:
https://drive.google.com/file/d/1evR-AKeqjRE_9-amH88ab8wB4e46xpEb/view?usp=share_link

Simply download the file and double click it with your Unity project already open.